wip: move connectivity_plus_linux into the main package
