+++ packages/connectivity_plus/connectivity_plus/lib/src/connectivity_plus_linux.dart
