+++ packages/connectivity_plus/connectivity_plus/pubspec.yaml
@@ -18,7 +18,8 @@ flutter:
       ios:
         pluginClass: ConnectivityPlusPlugin
       linux:
-        default_package: connectivity_plus_linux
+        dartPluginClass: ConnectivityLinux
+        pluginClass: none
       macos:
         default_package: connectivity_plus_macos
       web:
@@ -30,7 +31,6 @@ dependencies:
   flutter:
     sdk: flutter
   connectivity_plus_platform_interface: ^1.2.0
-  connectivity_plus_linux: ^1.3.0
   connectivity_plus_macos: ^1.2.4
   connectivity_plus_web: ^1.2.5
   connectivity_plus_windows: ^1.2.2
@@ -38,7 +38,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test: ^1.21.1
+  build_runner: ^2.1.11
+  dbus: ^0.7.5
   mockito: ^5.2.0
   plugin_platform_interface: ^2.1.2
-  flutter_lints: ^2.0.1
+  test: ^1.21.1
