+++ packages/connectivity_plus/connectivity_plus/test/connectivity_plus_linux_test.dart
@@ -1,4 +1,4 @@
-import 'package:connectivity_plus_linux/src/connectivity.dart';
+import 'package:connectivity_plus/src/connectivity_plus_linux.dart';
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
@@ -6,6 +6,7 @@ import 'package:mockito/mockito.dart';
 import 'package:nm/nm.dart';
 
 import 'connectivity_plus_linux_test.mocks.dart';
+
 
 @GenerateMocks([NetworkManagerClient])
 void main() {
