--- packages/connectivity_plus/connectivity_plus_linux/lib/connectivity_plus_linux.dart
@@ -1,4 +1,0 @@
-/// The Linux implementation of `connectivity_plus`.
-library connectivity_plus_linux;
-
-export 'src/connectivity.dart';
